refactor(cli): split profile command facade
